Refactor project structure and clean main
Reorganized project files into clearer modules (Rooms, EnemyGenerator, UI, Difficulty)

Simplified main.cpp

Added enemy generation system based on room and difficulty

Improved UI handling for enemy messages

Fixed several logic bugs in combat flow and enemy generation

+++ Project2/Inventario.cpp
@@ -6,7 +6,7 @@
 #include <chrono>
 #include <thread>
 
-int Inventario::getQuantidadeItens() const {
+size_t Inventario::getQuantidadeItens() const {
     return itens.size();
 }
